Refactor HitDice into gamesystem.
