+++ gamesystem/CreatureProcessor.java
@@ -3,7 +3,6 @@ package gamesystem;
 import gamesystem.XP.XPChangeAdhoc;
 import gamesystem.XP.XPChangeChallenges;
 import gamesystem.XP.XPChangeLevel;
-import monsters.HitDice;
 import monsters.Monster.MonsterAttackRoutine;
 import party.CharacterAttackForm;
